+++ storage/pebble/storage_test.go
@@ -1,7 +1,7 @@
 package pebble
 
 import (
-	"github.com/onflow/flow-evm-gateway/storage"
+	"github.com/onflow/flow-evm-gateway/storage/mocks"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
@@ -12,7 +12,7 @@ func TestBlock(t *testing.T) {
 		db, err := New()
 		require.NoError(t, err)
 
-		block := storage.NewBlock(10)
+		block := mocks.NewBlock(10)
 		db.storeBlock()
 	})
 }
